Changed TrackDtos .PitBoxesCount, .ServerSlotsCount property type from ushort to byte.

+++ Models/DTOs/Individual/Track.cs
@@ -24,8 +24,8 @@ namespace GTRC_Basics.Models.DTOs
     {
         public string? Name { get; set; }
         public string? AccTrackId { get; set; }
-        public ushort? PitBoxesCount { get; set; }
-        public ushort? ServerSlotsCount { get; set; }
+        public byte? PitBoxesCount { get; set; }
+        public byte? ServerSlotsCount { get; set; }
         public ushort? AccTimePenDtS { get; set; }
         public string? NameGoogleSheets { get; set; }
     }
